+++ src/components/Logout.jsx
@@ -1,4 +1,4 @@
-import { auth } from '../../firebase';
+import { auth } from '../firebase';
 import { signOut } from 'firebase/auth';
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
